Registration and login very close
The site has functional registration and a user can very nearly log in.  The  server returns separate responses for a valid user and a non-valid one, now I just have to have it properly return the user object and store it appropriately so I can access it to store products on.