There's a lot here
Ok so yeah, this is too much for one commit but sometimes you don't know when something is going to be baseline functional or not.

So.  Users can now register and login to the system, with said user object stored in local storage.  I'll add things like ssl and storing the object more securely but right now it works.

I've done some playing around with scrapers and tried to add that.  It kinda works? But I now have more duplication in my db.  New imported products can be viewed on the site though.

I've also made some changes to how products are displayed.  After some fiddling I have a non-masonry grid of images displayed, with the starts of adding category-specific classes to the product by brand page.  Once I know what I'm doing, I'll update these classes to use sass variables, or a mixin, depends.  I just wanted to get a better grid of images working so I have a better base from which to build the collection of products.