Verified functionality on new computer and fixed brand list
Since getting a new laptop, there were some weird inconsistencies that come with system level settings for running a Node server.  It now runs properly and all the previous functionality has been confirmed.

On the brand list page, the sort feature wasn't actually functional so I built that in and updated the links so that it works.  Also added some more formatting to the product page but that needs woooork.

Onwards!